#293 Switch to integer sizes for diagram variables.
- Floating point gives too much trouble for pixel-perfect connections.
- Rectangle marcher test uses new inner and outer radius.
  This changes the affected set pixels and coverage of pixels.
- Use a Position2D to pass the center point into Marching Rectangles.
- Add a relative Position2D for specifying the starting pixel with MarchingRectangles.
- Handle empty graphics in the solver by allowing -1 as value for a variable.

+++ common/org.eclipse.escet.common.raildiagrams/src/org/eclipse/escet/common/raildiagrams/graphics/Area.java
@@ -60,10 +60,9 @@ public abstract class Area {
      * @param xOffset Horizontal offset of the element in the picture.
      * @param yOffset Vertical offset of the element in the picture.
      */
-    public void dump(Solver solver, double xOffset, double yOffset) {
-        dbg("%s: x[%.1f--%.1f], y[%.1f--%.1f]", prefix, xOffset + solver.getVarValue(left),
-                xOffset + solver.getVarValue(right), yOffset + solver.getVarValue(top),
-                yOffset + solver.getVarValue(bottom));
+    public void dump(Solver solver, int xOffset, int yOffset) {
+        dbg("%s: x[%d--%d], y[%d--%d]", prefix, xOffset + solver.getVarValue(left), xOffset + solver.getVarValue(right),
+                yOffset + solver.getVarValue(top), yOffset + solver.getVarValue(bottom));
     }
 
     /**
@@ -74,5 +73,5 @@ public abstract class Area {
      * @param solver Solver containing values for all variables.
      * @return Positions expected to be used by other elements.
      */
-    public abstract Position2D[] getConnectPoints(double baseLeft, double baseTop, Solver solver);
+    public abstract Position2D[] getConnectPoints(int baseLeft, int baseTop, Solver solver);
 }
